Move `Worker` to `execution`
